Use true relative links
The `relative_url` liquid filter generates
domain-relative absolute paths, not relative paths.
https://jekyllrb.com/docs/liquid/filters/
https://github.com/jekyll/jekyll/issues/6360

That was breaking the site when it was proxied to w3.org/2020/maps/

This switches to a template-based solution
that calculates a relative URL from the current page
back to the repo root (like ../../ etc)
and prepends that to the file paths.
Code from https://gist.github.com/Ketrel/1f36fca6b4148f7263b3ee8fdb0923e7
(using the variation by Sakrecoer)

Note that I've included it on all the content pages,
in addition to the default layout template,
since content is processed before the template.
Alternatively, the content pages could just be hard-coded
with relative URLs, based on the permalink for that page.

This works for everything except the 404 page,
since the page.url when generating the 404 is the file path,
not the URL that 404'd.
So, the 404 page has been replaced by a bare-bones page
with a hard-coded link to the site URL.