Move the position statements folder
From the “assets” folder to a new “supporting-material-uploads” folder.

Couldn't name the folder “supporting-material”,
because that caused problems with URL handling,
because the /supporting-material URL is *not*
inside the folder, it's an alias for /supporting-material.html
Although Jekyll served the same file eitherway,
the relative links were all broken!

(I had to clear the cach to get Chrome to forget that this URL was ever a folder.
The Jekyll localhost must have sent a permanent redirect header.)