Reformat agenda page for Video & discussion links
We're only posting one video per day,
so put it up next to the heading for the day,
with extra formatting to call it out.

De-emphasize the session time when a video is available.

Add a discourse link to individual talks.

Support alternatives to “slides” link & suppress the (PDF) on them.

Tweak some styling (of course).

Add the links we have, plus a background-reading link for the COGA panel
(I gave up on making a template for that. Back to HTML in the YAML.)