Draft report - More copy edits
This finishes my review of the text.

I've removed the "Editor's note" asides (I assume from @prushforth?), as they didn't really seem part of a review of the workshop. If they could be re-written as cross-references to information presented in a different session, that would be one thing, but interjections from an anonymous editor don't seem appropriate to me.

Other than that, mostly editing / re-organizing for readability.